# Covenant/API/Models/ReferenceAssembly.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Covenant.API.Models
{
    using Newtonsoft.Json;
    using System.Collections;
    using System.Collections.Generic;
    using System.Linq;

    public partial class ReferenceAssembly
    {
        /// <summary>
        /// Initializes a new instance of the ReferenceAssembly class.
        /// </summary>
        public ReferenceAssembly()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the ReferenceAssembly class.
        /// </summary>
        /// <param name="dotNetVersion">Possible values include: 'Net40',
        /// 'Net35', 'NetCore21', 'NetCore30'</param>
        public ReferenceAssembly(int? id = default(int?), string name = default(string), string location = default(string), DotNetVersion? dotNetVersion = default(DotNetVersion?), IList<ReferenceSourceLibrary> referenceSourceLibraries = default(IList<ReferenceSourceLibrary>), IList<GruntTask> gruntTasks = default(IList<GruntTask>))
        {
            Id = id;
            Name = name;
            Location = location;
            DotNetVersion = dotNetVersion;
            ReferenceSourceLibraries = referenceSourceLibraries;
            GruntTasks = gruntTasks;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public int? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "name")]
        public string Name { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "location")]
        public string Location { get; set; }

        /// <summary>
        /// Gets or sets possible values include: 'Net40', 'Net35',
        /// 'NetCore21', 'NetCore30'
        /// </summary>
        [JsonProperty(PropertyName = "dotNetVersion")]
        public DotNetVersion? DotNetVersion { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "referenceSourceLibraries")]
        public IList<ReferenceSourceLibrary> ReferenceSourceLibraries { get; private set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "gruntTasks")]
        public IList<GruntTask> GruntTasks { get; private set; }

    }
}