# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Covenant API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/listeners/{id}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetListener",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Listener"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "DeleteListener",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/listeners/http/{id}": {
      "get": {
        "tags": [
          "ListenerApi"
        ],
        "operationId": "GetActiveHttpListener",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpListener"
            }
          }
        }
      }
    },
...
    "/api/profiles/http/{id}": {
      "get": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "GetHttpProfile",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpProfile"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "DeleteHttpProfile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/referenceassemblies": {
      "get": {
        "tags": [
          "ReferenceAssemblyApi"
        ],
        "operationId": "GetReferenceAssemblies",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ReferenceAssembly"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReferenceAssemblyApi"
        ],
        "operationId": "CreateReferenceAssembly",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReferenceAssembly"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReferenceAssembly"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ReferenceAssemblyApi"
        ],
        "operationId": "EditReferenceAssembly",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReferenceAssembly"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReferenceAssembly"
            }
          }
        }
      }
    },
    "/api/referenceassemblies/{id}": {
      "get": {
        "tags": [
          "ReferenceAssemblyApi"
        ],
        "operationId": "GetReferenceAssembly",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReferenceAssembly"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ReferenceAssemblyApi"
        ],
        "operationId": "DeleteReferenceAssembly",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DotNetVersion": {
      "enum": [
        "Net40",
        "Net35",
        "NetCore21",
        "NetCore30"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "DotNetVersion",
        "modelAsString": false
      }
    },
    "ImplantLanguage": {
      "enum": [
        "CSharp"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantLanguage",
        "modelAsString": false
      }
    },
    "GruntTaskingType": {
      "enum": [
        "Assembly",
        "SetOption",
        "Exit",
        "Connect",
        "Disconnect",
        "Jobs"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "GruntTaskingType",
        "modelAsString": false
      }
    },
    "GruntTaskOption": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "defaultValue": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "suggestedValues": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "optional": {
          "type": "boolean"
        },
        "displayInCommand": {
          "type": "boolean"
        },
        "gruntTaskId": {
          "format": "int32",
          "type": "integer"
        },
        "task": {
          "$ref": "#/definitions/GruntTask"
        }
      }
    },
    "GruntTask": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "alternateNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": {
          "type": "string"
        },
        "help": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "supportedDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "code": {
          "type": "string"
        },
        "taskingType": {
          "$ref": "#/definitions/GruntTaskingType"
        },
        "compiled": {
          "type": "boolean"
        },
        "unsafeCompile": {
          "type": "boolean"
        },
        "tokenTask": {
          "type": "boolean"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GruntTaskOption"
          }
        },
        "referenceSourceLibraries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceSourceLibrary"
          },
          "readOnly": true
        },
        "referenceAssemblies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceAssembly"
          },
          "readOnly": true
        },
        "embeddedResources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmbeddedResource"
          },
          "readOnly": true
        }
      }
    },
    "ReferenceAssembly": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "dotNetVersion": {
          "$ref": "#/definitions/DotNetVersion"
        },
        "referenceSourceLibraries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceSourceLibrary"
          },
          "readOnly": true
        },
        "gruntTasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GruntTask"
          },
          "readOnly": true
        }
      }
    },
    "ReferenceSourceLibrary": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "supportedDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "referenceAssemblies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceAssembly"
          },
          "readOnly": true
        },
        "embeddedResources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmbeddedResource"
          },
          "readOnly": true
        },
        "gruntTasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GruntTask"
          },
          "readOnly": true
        }
      }
    },
    "EmbeddedResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "referenceSourceLibraries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceSourceLibrary"
          },
          "readOnly": true
        },
        "gruntTasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GruntTask"
          },
          "readOnly": true
        }
      }
    },
...
    "CommunicationType": {
      "enum": [
        "HTTP",
        "SMB",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "CommunicationType",
        "modelAsString": false
      }
    },
    "ImplantDirection": {
      "enum": [
        "Push",
        "Pull"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantDirection",
        "modelAsString": false
      }
    },
    "ListenerType": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "compatibleImplantTemplates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImplantTemplate"
          },
          "readOnly": true
        }
      }
    },
    "ImplantTemplate": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "commType": {
          "$ref": "#/definitions/CommunicationType"
        },
        "implantDirection": {
          "$ref": "#/definitions/ImplantDirection"
        },
        "compatibleListenerTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListenerType"
          },
          "readOnly": true
        },
        "stagerCode": {
          "type": "string"
        },
        "executorCode": {
          "type": "string"
        }
      }
    },
    "ProfileType": {
      "enum": [
        "HTTP",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ProfileType",
        "modelAsString": false
      }
    },
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/ProfileType"
        },
        "messageTransform": {
          "type": "string"
        }
      }
    },
    "ListenerStatus": {
      "enum": [
        "Uninitialized",
        "Active",
        "Stopped"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ListenerStatus",
        "modelAsString": false
      }
    },
    "Listener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "guid",
        "listenerTypeId",
        "name",
        "profileId",
        "status"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "guid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "HttpListener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "guid",
        "listenerTypeId",
        "name",
        "profileId",
        "status",
        "urls",
        "useSSL"
      ],
      "type": "object",
      "properties": {
        "useSSL": {
          "type": "boolean"
        },
        "sslCertificate": {
          "type": "string"
        },
        "sslCertificatePassword": {
          "type": "string"
        },
        "sslCertHash": {
          "type": "string"
        },
        "urls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "guid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "HttpProfileHeader": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "HttpProfile": {
      "type": "object",
      "properties": {
        "httpUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "httpRequestHeaders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HttpProfileHeader"
          }
        },
        "httpResponseHeaders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HttpProfileHeader"
          }
        },
        "httpPostRequest": {
          "type": "string"
        },
        "httpGetResponse": {
          "type": "string"
        },
        "httpPostResponse": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/ProfileType"
        },
        "messageTransform": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  }
}